refactor: extract storage layout module
